refactor(server): extract shared TLS test utilities
